fixed edge case that sometimes an "old" frame or a "too new" frame gets shown

+++ include/DMDUtil/DMD.h
@@ -8,7 +8,7 @@
 #define DMDUTILCALLBACK
 #endif
 
-#define DMDUTIL_FRAME_BUFFER_SIZE 64
+#define DMDUTIL_FRAME_BUFFER_SIZE 64 // Must be a divider of 65535!
 #define DMDUTIL_MIN_FRAMES_BEHIND 4
 #define DMDUTIL_MAX_FRAMES_BEHIND 32
 #define DMDUTIL_MAX_NAME_SIZE 16
@@ -186,7 +186,7 @@ class DMDUTILAPI DMD
   Update* m_pUpdateBufferQueue[DMDUTIL_FRAME_BUFFER_SIZE];
   std::shared_ptr<Update> m_updateBuffered;
 
-  uint8_t GetNextBufferQueuePosition(uint8_t bufferPosition, const uint8_t updateBufferQueuePosition);
+  uint8_t GetNextBufferQueuePosition(uint16_t bufferPosition, const uint16_t updateBufferQueuePosition);
   bool ConnectDMDServer();
   bool UpdatePalette(uint8_t* pPalette, uint8_t depth, uint8_t r, uint8_t g, uint8_t b);
   void UpdateData(const uint8_t* pData, int depth, uint16_t width, uint16_t height, uint8_t r, uint8_t g, uint8_t b,
@@ -234,7 +234,7 @@ class DMDUTILAPI DMD
   std::condition_variable_any m_dmdCV;
   std::atomic<bool> m_dmdFrameReady;
   std::atomic<bool> m_stopFlag;
-  std::atomic<uint8_t> m_updateBufferQueuePosition;
+  std::atomic<uint16_t> m_updateBufferQueuePosition;
   std::atomic<uint16_t> m_pupSceneId;
 
   bool m_hasUpdateBuffered = false;
